+++ src/main/java/backend/ApiImpl.java
@@ -1,12 +1,14 @@
 package backend;
 
 import database.Database;
-import database.MockDatabase;
+import java.io.IOException;
 import java.util.List;
 import protocol.LobbyModel;
+import protocol.ProgressSnapshot;
 import protocol.Response;
 import protocol.ResponseFactory;
 import server.PushService;
+import util.Logger;
 
 
 /**
@@ -15,53 +17,69 @@ import server.PushService;
 public class ApiImpl implements Api {
 
   private final PushService pushService;
-  private final LobbyStore lobbyStore;
+  private final SessionStore sessionStore;
   private final Database database;
+  private final RaceSettings raceSettings;
 
   /**
    * Create an instance implementing the Api interface.
    *
    * @param pushService service where responses are sent to
    */
-  public ApiImpl(PushService pushService) {
+  public ApiImpl(PushService pushService, Database database) {
     this.pushService = pushService;
-    this.lobbyStore = new LobbyStore();
-    this.database = new MockDatabase();
+    this.database = database;
+    this.sessionStore = new SessionStore(this.database);
+    this.raceSettings = new RaceSettings(5, 1);
   }
 
   @Override
   public void registerPlayer(String connectionId, String name) {
-    String userId = database.registerUser(name);
-    Response response = ResponseFactory.makeRegisteredResponse(userId);
-    pushService.sendResponse(connectionId, response);
+    try {
+      String userId = database.registerUser(name);
+      Response response = ResponseFactory.makeRegisteredResponse(userId);
+      pushService.sendResponse(connectionId, response);
+    } catch (IOException e) {
+      Logger.logError(e.getMessage());
+    }
   }
 
   @Override
   public void createNewLobby(String connectionId, String userId) {
-    String lobbyId = lobbyStore.createNewLobby(connectionId, pushService);
-    lobbyStore.joinLobby(lobbyId, connectionId, userId);
+    String lobbyId = sessionStore.createNewLobby(connectionId, pushService);
+    sessionStore.joinLobby(lobbyId, connectionId, userId);
   }
 
   @Override
   public void joinLobby(String lobbyId, String connectionId, String userId) {
-    lobbyStore.joinLobby(lobbyId, connectionId, userId);
+    sessionStore.joinLobby(lobbyId, connectionId, userId);
   }
 
   @Override
   public void leaveLobby(String connectionId) {
-    lobbyStore.leaveLobby(connectionId);
+    sessionStore.leaveLobby(connectionId);
   }
 
   @Override
   public void startRace(String connectionId) {
-    lobbyStore.startGame(connectionId);
+    sessionStore.startGame(connectionId, raceSettings);
   }
 
   @Override
   public void getLobbies(String connectionId) {
-    List<LobbyModel> lobbies = lobbyStore.getOpenLobbies();
+    List<LobbyModel> lobbies = sessionStore.getOpenLobbies();
     Response response = ResponseFactory.makeLobbiesResponse(lobbies);
     pushService.sendResponse(connectionId, response);
+  }
+
+  @Override
+  public void setPlayerReady(String connectionId, boolean isReady) {
+    sessionStore.setPlayerReady(connectionId, isReady);
+  }
+
+  @Override
+  public void updateProgress(String connectionId, ProgressSnapshot snapshot) {
+    sessionStore.updateProgress(connectionId, snapshot);
   }
 
 }