+++ src/main/java/client/Main.java
@@ -2,8 +2,11 @@ package client;
 
 import java.io.IOException;
 import java.net.InetAddress;
+import java.time.Instant;
 import java.util.List;
 import protocol.LobbyModel;
+import protocol.PlayerUpdate;
+import protocol.ProgressSnapshot;
 import protocol.RaceModel;
 
 /**
@@ -24,9 +27,17 @@ public class Main implements ClientObserver {
       client.subscribe(m);
       client.registerUser("Cooler Mensch");
       client.requestLobbies();
-      client.newGame("some userId");
+      client.newLobby("some userId");
       client.requestLobbies();
-      client.startGame();
+      client.setIsReady(true);
+      client.startRace();
+      client.sendProgressUpdate(new ProgressSnapshot(0, 30, 5, 0));
+      try {
+        Thread.sleep(5000);
+      } catch (InterruptedException e) {
+        e.printStackTrace();
+      }
+      client.sendProgressUpdate(new ProgressSnapshot(0, 60, 10, 0));
     } catch (IOException e) {
       System.out.print(e.getMessage());
     }
@@ -49,17 +60,27 @@ public class Main implements ClientObserver {
 
   @Override
   public void receivedLobbyUpdate(LobbyModel lobby) {
-    System.out.println("Received Lobby update:");
-    System.out.println(lobby.players);
-    System.out.println(lobby.id);
+    System.out.println("Received Lobby update: " + lobby);
     System.out.println(lobby.isRunning);
+    System.out.println(lobby.id);
+    System.out.println(lobby.players);
   }
 
   @Override
   public void receivedOpenLobbies(List<LobbyModel> lobbies) {
-    System.out.println("Open lobbies:");
-    System.out.println(lobbies);
+    System.out.println("Open lobbies: " + lobbies);
   }
 
+  @Override
+  public void receivedRaceUpdate(List<PlayerUpdate> updates) {
+    System.out.println("Race update: " + updates);
+    System.out.println("Wpm: " + updates.get(0).wpm);
+    System.out.println("Wpm: " + updates.get(0).percentProgress);
+  }
+
+  @Override
+  public void receivedCheckeredFlag(long raceStop) {
+    System.out.println("Checkered Flag! Race stopped: " + Instant.ofEpochSecond(raceStop));
+  }
 
 }