+++ src/main/java/model/GamePhase.java
@@ -1,0 +1,7 @@
+package model;
+
+/** Enum class for game phases. */
+public enum GamePhase {
+  RUNNING,
+  FINISHED,
+}