+++ src/main/java/protocol/RaceModel.java
@@ -1,5 +1,7 @@
 package protocol;
 
+import java.util.List;
+
 /**
  * Json model including all race information.
  */
@@ -7,13 +9,16 @@ public class RaceModel {
 
   public final String textToType;
 
+  public final List<String> players;
+
   /**
    * Create RaceModel.
    *
    * @param textToType text of the race
    */
-  public RaceModel(String textToType) {
+  public RaceModel(String textToType, List<String> players) {
     this.textToType = textToType;
+    this.players = players;
   }
 
 }